Revert "Reformula organizacao"
