+++ .gitignore
@@ -26,5 +26,3 @@ hs_err_pid*
 .project
 .settings/*
 .*
-/nbproject/private/
-/build/
