--- nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=83a560af
-build.xml.script.CRC32=7d980f80
-build.xml.stylesheet.CRC32=f85dc8f2@1.93.0.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=83a560af
-nbproject/build-impl.xml.script.CRC32=7380f2b6
-nbproject/build-impl.xml.stylesheet.CRC32=f89f7d21@1.93.0.48
