--- src/principal/Main.java
@@ -1,31 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package principal;
-
-import expressoes.Posfixa;
-import java.util.Scanner;
-
-/**
- *
- * @author anderson
- */
-public class Main 
-{
-    public static void main(String[] args)
-    {
-	Scanner scanner = new Scanner(System.in);
-	String expressao = "";
-	
-	if(args.length == 0)
-	{
-            System.out.println("Informe a expressão infixa: ");
-            expressao = scanner.nextLine().replace(" ", "");
-	}
-	else
-            expressao = args[0];
-	System.out.println(Posfixa.conversao(expressao));
-    }
-}
